server/requirements.txt
fastapi==0.118.0
uvicorn==0.30.6
google-generativeai==0.7.2
python-dotenv==1.0.1